renamed commands module to cli
