+++ gens/cli/view.py
